Added ability to specify patient widgets configs
