+++ src/config-schemas/openmrs-esm-patient-chart-schema.ts
@@ -3,7 +3,12 @@ import {
   coreWidgetDefinitions,
   coreDashboardDefinitions,
   coreTabbedViewDefinitions
-} from "./view-components/core-views";
+} from "../view-components/core-views";
+
+import {
+  ChartWidgetsConfig,
+  chartWidgetsConfigSchemas
+} from "@openmrs/esm-patient-chart-widgets";
 
 export const esmPatientChartSchema = {
   primaryNavbar: {
@@ -99,5 +104,40 @@ export const esmPatientChartSchema = {
       }
     },
     default: coreTabbedViewDefinitions
-  }
+  },
+  ...chartWidgetsConfigSchemas
 };
+
+export type ChartConfig = ChartWidgetsConfig & {
+  primaryNavbar: Navbar[];
+  widgetDefinitions: {
+    name: string;
+    esModule?: string;
+    label?: string;
+    path?: string;
+  };
+
+  dashboardDefinitions: {
+    name: string;
+    title: string;
+    layout: { columns: number };
+    widgets: {
+      name: string;
+      esModule: string;
+      label: string;
+      path: string;
+      layout: {
+        rowSpan: number;
+        columnSpan: number;
+      };
+    }[];
+  };
+
+  tabbedDashboardDefinitions: {
+    name: string;
+    title: string;
+    navbar: Navbar;
+  }[];
+};
+
+export type Navbar = { label: string; path: string; view: string };
